Added seperate operand class files
