+++ tests/test_parser_AArch64.py
@@ -8,7 +8,7 @@ import unittest
 
 from pyparsing import ParseException
 
-from osaca.parser import AttrDict, ParserAArch64
+from osaca.parser import AttrDict, ParserAArch64, InstructionForm
 
 
 class TestParserAArch64(unittest.TestCase):
@@ -458,19 +458,13 @@ class TestParserAArch64(unittest.TestCase):
         )
 
     def _get_label(self, parser, label):
-        return AttrDict.convert_dict(
-            parser.process_operand(parser.label.parseString(label, parseAll=True).asDict())
-        ).label
+        return parser.process_operand(parser.label.parseString(label, parseAll=True).asDict()).label
 
     def _get_directive(self, parser, directive):
-        return AttrDict.convert_dict(
-            parser.process_operand(parser.directive.parseString(directive, parseAll=True).asDict())
-        ).directive
+        return parser.process_operand(parser.directive.parseString(directive, parseAll=True).asDict()).directive
 
     def _get_condition(self, parser, condition):
-        return AttrDict.convert_dict(
-            parser.process_operand(parser.condition.parseString(condition, parseAll=True).asDict())
-        ).condition
+        return parser.process_operand(parser.condition.parseString(condition, parseAll=True).asDict()).condition
 
     @staticmethod
     def _find_file(name):
